+++ tests/test_uri_templates.py
@@ -13,6 +13,7 @@ import six
 
 import falcon
 from falcon import testing
+from falcon.routing.util import SuffixedMethodNotFoundError
 
 
 _TEST_UUID = uuid.uuid4()
@@ -80,6 +81,40 @@ class FileDetailsResource(object):
         self.called = True
 
 
+class ResourceWithSuffixRoutes(object):
+    def __init__(self):
+        self.get_called = False
+        self.post_called = False
+        self.put_called = False
+
+    def on_get(self, req, resp, collection_id, item_id):
+        self.collection_id = collection_id
+        self.item_id = item_id
+        self.get_called = True
+
+    def on_post(self, req, resp, collection_id, item_id):
+        self.collection_id = collection_id
+        self.item_id = item_id
+        self.post_called = True
+
+    def on_put(self, req, resp, collection_id, item_id):
+        self.collection_id = collection_id
+        self.item_id = item_id
+        self.put_called = True
+
+    def on_get_collection(self, req, resp, collection_id):
+        self.collection_id = collection_id
+        self.get_called = True
+
+    def on_post_collection(self, req, resp, collection_id):
+        self.collection_id = collection_id
+        self.post_called = True
+
+    def on_put_collection(self, req, resp, collection_id):
+        self.collection_id = collection_id
+        self.put_called = True
+
+
 @pytest.fixture
 def resource():
     return testing.SimpleTestResource()
@@ -132,7 +167,7 @@ def test_special_chars(client, resource):
     'widget_id',
 ])
 def test_single(client, resource, field_name):
-    template = '/widgets/{{{0}}}'.format(field_name)
+    template = '/widgets/{{{}}}'.format(field_name)
 
     client.app.add_route(template, resource)
 
@@ -224,17 +259,17 @@ def test_datetime_converter(client, resource, uri_template, path, dt_expected):
     ),
     (
         '/versions/diff/{left:uuid()}...{right:uuid()}',
-        '/versions/diff/{0}...{1}'.format(_TEST_UUID_STR, _TEST_UUID_STR_2),
+        '/versions/diff/{}...{}'.format(_TEST_UUID_STR, _TEST_UUID_STR_2),
         {'left': _TEST_UUID, 'right': _TEST_UUID_2, }
     ),
     (
         '/versions/diff/{left:uuid}...{right:uuid()}',
-        '/versions/diff/{0}...{1}'.format(_TEST_UUID_STR, _TEST_UUID_STR_2),
+        '/versions/diff/{}...{}'.format(_TEST_UUID_STR, _TEST_UUID_STR_2),
         {'left': _TEST_UUID, 'right': _TEST_UUID_2, }
     ),
     (
         '/versions/diff/{left:uuid()}...{right:uuid}',
-        '/versions/diff/{0}...{1}'.format(_TEST_UUID_STR, _TEST_UUID_STR_2),
+        '/versions/diff/{}...{}'.format(_TEST_UUID_STR, _TEST_UUID_STR_2),
         {'left': _TEST_UUID, 'right': _TEST_UUID_2, }
     ),
     (
@@ -263,7 +298,7 @@ def test_uuid_converter_complex_segment(client, resource):
     first_uuid = uuid.uuid4()
     last_uuid = uuid.uuid4()
 
-    result = client.simulate_get('/pages/{0}...{1}'.format(
+    result = client.simulate_get('/pages/{}...{}'.format(
         first_uuid,
         last_uuid
     ))
@@ -399,3 +434,39 @@ def test_same_level_complex_var(client, reverse):
     assert details_resource.called
     assert details_resource.file_id == file_id_2
     assert details_resource.ext == ext
+
+
+def test_adding_suffix_routes(client):
+    resource_with_suffix_routes = ResourceWithSuffixRoutes()
+    client.app.add_route(
+        '/collections/{collection_id}/items/{item_id}', resource_with_suffix_routes)
+    client.app.add_route(
+        '/collections/{collection_id}/items', resource_with_suffix_routes, suffix='collection')
+    # GET
+    client.simulate_get('/collections/123/items/456')
+    assert resource_with_suffix_routes.collection_id == '123'
+    assert resource_with_suffix_routes.item_id == '456'
+    assert resource_with_suffix_routes.get_called
+    client.simulate_get('/collections/foo/items')
+    assert resource_with_suffix_routes.collection_id == 'foo'
+    # POST
+    client.simulate_post('/collections/foo234/items/foo456')
+    assert resource_with_suffix_routes.collection_id == 'foo234'
+    assert resource_with_suffix_routes.item_id == 'foo456'
+    assert resource_with_suffix_routes.post_called
+    client.simulate_post('/collections/foo123/items')
+    assert resource_with_suffix_routes.collection_id == 'foo123'
+    # PUT
+    client.simulate_put('/collections/foo345/items/foo567')
+    assert resource_with_suffix_routes.collection_id == 'foo345'
+    assert resource_with_suffix_routes.item_id == 'foo567'
+    assert resource_with_suffix_routes.put_called
+    client.simulate_put('/collections/foo321/items')
+    assert resource_with_suffix_routes.collection_id == 'foo321'
+
+
+def test_custom_error_on_suffix_route_not_found(client):
+    resource_with_suffix_routes = ResourceWithSuffixRoutes()
+    with pytest.raises(SuffixedMethodNotFoundError):
+        client.app.add_route(
+            '/collections/{collection_id}/items', resource_with_suffix_routes, suffix='bad-alt')