+++ CHANGES.rst
@@ -1,10 +1,55 @@
-1.4.0
+2.0.0
+=====
+
+Breaking Changes
+----------------
+
+- ``testing.Result.json`` now returns ``None`` when the response body is
+  empty, rather than raising an error.
+
+Changes to Supported Platforms
+------------------------------
+
+New & Improved
+--------------
+
+Fixed
+-----
+
+1.4.1
 =====
 
 Breaking Changes
 ----------------
 
 (None)
+
+Changes to Supported Platforms
+------------------------------
+
+(None)
+
+New & Improved
+--------------
+
+(None)
+
+Fixed
+-----
+
+- Reverted the breaking change in 1.4.0 to ``falcon.testing.Result.json``.
+  Minor releases should have no breaking changes.
+- The README was not rendering properly on PyPI. This was fixed and a validation
+  step was added to the build process.
+
+1.4.0
+=====
+
+Breaking Changes
+----------------
+
+- ``testing.Result.json`` now returns ``None`` when the response body is
+  empty, rather than raising an error.
 
 Changes to Supported Platforms
 ------------------------------
@@ -47,8 +92,6 @@ New & Improved
   override ``TestClient``'s behavior.
 - ``TestClient`` can now be configured with a default set of headers to
   send with every request.
-- ``testing.Result.json`` now returns None when the response body is
-  empty, rather than raising an error.
 - The FAQ has been reorganized and greatly expanded.
 - We restyled the docs to match https://falconframework.org